protocol/workspace: Move `WorkspaceGroupData` to `ext.rs`
More consistent to have this here next to `WorkspaceData`, now it isn't
shared with cosmic workspace v1.
